refactor: separate src files and add ts path aliases
This separates the TS files into `src/core` and `src/scripts` to
differentiate between the files that are used in the main app and the
standalone scripts that are run from the command line. On top of that,
we will have a `src/config` folder for the config files. Since the
files are ignored by git, we need to add a `.gitkeep` file to keep the
folder in the repo.

Lastly, to make the imports easier, we will add path aliases to the
`tsconfig.json` file so we can drop the relative paths
