+++ tsconfig.json
@@ -1,6 +1,12 @@
 {
   "extends": "@tsconfig/node18/tsconfig.json",
   "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@core/*": ["src/core/*"],
+      "@config/*": ["src/config/*"],
+      "@scripts/*": ["src/scripts/*"],
+    },
     "allowJs": true,
     "removeComments": true, /* Remove comments */
     "sourceMap": true,  /* Generates corresponding '.map' file. */
@@ -8,5 +14,4 @@
     "declaration": true, /* Generates corresponding '.d.ts' file. */
     "noImplicitAny": false,
   },
-  "include" : ["src/**/*", "lh-config.js"]
 }
